+++ src/pages/NewsPage.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
-
+import News from '../components/News/index'
 const NewsPage = () => {
   return <div style={{paddingTop:100}}>
-    <h3>News</h3>
+    <h3>Latest News</h3>
+    <News/>
   </div>;
 };