+++ changelog.md
@@ -1,5 +1,13 @@
 
 
+[0.0.2] 2022-02-01
+-`added`
+- src > components > News components
+-`installed`
+- npm install @reduxjs/toolkit react-redux
+-`fixed`
+- search bar error(changed the initial value of 'search' to "")
+
 [0.0.1] 2022-01-31
 
 -`added`
@@ -11,5 +19,3 @@
 -`installed`
 - react-creat-app 
 - installed @mui/material styled-components @mui/icons-material react-router-dom
-
--`changed`